requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
autotrain-advanced
requests
django-cors-headers
fastapi